# Conflicts:
#	.github/dependabot.yml
#	client/css/tailwind.css
#	client/fe/webpack/moduleRules.js
#	client/fe/webpack/resolveAliases.js
#	client/js/InitVue.js
#	client/js/VueConfigLocal.js
#	client/js/bundles/procedure/institutionTagManagement.js
#	client/js/bundles/user/customersettingsmail.js
#	client/js/components/addon/AddonWrapper.vue
#	client/js/components/document/ElementsAdminList.vue
#	client/js/components/faq/DpFaqItem.vue
#	client/js/components/faq/DpFaqList.vue
#	client/js/components/map/admin/AdminLayerListItem.vue
#	client/js/components/map/publicdetail/Map.vue
#	client/js/components/map/publicdetail/controls/layerlist/DpPublicLayerListLayer.vue
#	client/js/components/map/publicdetail/controls/layerlist/DpPublicLayerListWrapper.vue
#	client/js/components/procedure/SegmentsList/CustomSearch.vue
#	client/js/components/procedure/SegmentsList/FilterFlyout.vue
#	client/js/components/procedure/SegmentsList/SegmentsList.vue
#	client/js/components/procedure/Shared/SimilarStatementSubmitters/SimilarStatementSubmitters.vue
#	client/js/components/procedure/StatementSegmentsList/SegmentsRecommendations.vue
#	client/js/components/procedure/StatementSegmentsList/StatementSegment.vue
#	client/js/components/procedure/StatementSegmentsList/StatementSegmentsEdit.vue
#	client/js/components/procedure/StatementSegmentsList/StatementSegmentsList.vue
#	client/js/components/procedure/admin/AdministrationPlaces.vue
#	client/js/components/procedure/admin/InstitutionTagManagement/InstitutionTagManagement.vue
#	client/js/components/procedure/basicSettings/AutoSwitchProcedurePhaseForm.vue
#	client/js/components/procedure/imageAnnotator/DpLabelModal.vue
#	client/js/components/procedure/publicindex/DpSearchProcedureMap.vue
#	client/js/components/shared/TextContentRenderer.vue
#	client/js/components/statement/DpBoilerPlateModal.vue
#	client/js/components/statement/DpStatementAnonymize.vue
#	client/js/components/statement/assessmentTable/AssessmentTableFilter.vue
#	client/js/components/statement/assessmentTable/ConsolidateModal.vue
#	client/js/components/statement/assessmentTable/DpAssessmentTableCard.vue
#	client/js/components/statement/assessmentTable/DpFilterModal.vue
#	client/js/components/statement/assessmentTable/DpMapModal.vue
#	client/js/components/statement/assessmentTable/ExportModal.vue
#	client/js/components/statement/assessmentTable/FilterModalSelectItem.vue
#	client/js/components/statement/assessmentTable/SearchModal/SearchModal.vue
#	client/js/components/statement/originalStatementsTable/OriginalStatementsTable.vue
#	client/js/components/statement/publicStatementLists/DpPublicStatement.vue
#	client/js/components/statement/publicStatementLists/DpPublicStatementList.vue
#	client/js/components/statement/statement/DpAutofillSubmitterData.vue
#	client/js/components/statement/statement/DpVersionHistory.vue
#	client/js/components/statement/statement/DpVersionHistoryDay.vue
#	client/js/components/statement/statement/StatementPublish.vue
#	client/js/components/statement/voter/StatementVoter.vue
#	client/js/components/user/CustomerSettings/CustomerSettingsSupport.vue
#	client/js/components/user/DpCreateItem.vue
#	client/js/components/user/DpOrganisationList/DpOrganisationFormFields.vue
#	client/js/components/user/DpOrganisationList/DpOrganisationListItem.vue
#	client/js/components/user/DpUserList/DpUserFormFields.vue
#	client/js/components/user/DpUserList/DpUserList.vue
#	client/js/store/core/initStore.js
#	client/js/store/map/Layers.js
#	client/js/store/procedure/SegmentFilter.js
#	client/js/store/statement/Fragment.js
#	client/js/store/statement/SplitStatementStore.js
#	config.webpack.js
#	package.json

+++ client/js/components/procedure/Shared/SimilarStatementSubmitters/SimilarStatementSubmitters.vue
@@ -16,145 +16,137 @@
       <dp-contextual-help :text="Translator.trans('statement.similarStatementSubmitters.hint')" />
     </div>
     <dp-editable-list
+      ref="listComponent"
+      class="o-list"
       :entries="listEntries"
       :has-permission-to-edit="editable"
       :translation-keys="translationKeys"
       @delete="deleteEntry"
       @reset="resetFormFields"
       @saveEntry="index => dpValidateAction('similarStatementSubmitterForm', () => handleSaveEntry(index), false)"
-      @show-update-form="setFormFields"
-      ref="listComponent">
+      @show-update-form="setFormFields">
       <template v-slot:list="{ entry, index }">
-        <ul class="o-list o-list--csv inline">
-          <template v-if="isRequestFormPost">
-            <input
-              type="hidden"
-              :name="'r_similarStatementSubmitters[' + index + '][fullName]'"
-              :value="entry.submitterName">
+        <template v-if="isRequestFormPost">
+          <input
+            type="hidden"
+            :name="'r_similarStatementSubmitters[' + index + '][fullName]'"
+            :value="entry.submitterName">
 
-            <input
-              type="hidden"
-              :name="'r_similarStatementSubmitters[' + index + '][city]'"
-              :value="entry.submitterCity">
+          <input
+            type="hidden"
+            :name="'r_similarStatementSubmitters[' + index + '][city]'"
+            :value="entry.submitterCity">
 
-            <input
-              type="hidden"
-              :name="'r_similarStatementSubmitters[' + index + '][streetName]'"
-              :value="entry.submitterAddress">
+          <input
+            type="hidden"
+            :name="'r_similarStatementSubmitters[' + index + '][streetName]'"
+            :value="entry.submitterAddress">
 
-            <input
-              type="hidden"
-              :name="'r_similarStatementSubmitters[' + index + '][streetNumber]'"
-              :value="entry.submitterHouseNumber">
+          <input
+            type="hidden"
+            :name="'r_similarStatementSubmitters[' + index + '][streetNumber]'"
+            :value="entry.submitterHouseNumber">
 
-            <input
-              type="hidden"
-              :name="'r_similarStatementSubmitters[' + index + '][postalCode]'"
-              :value="entry.submitterPostalCode">
+          <input
+            type="hidden"
+            :name="'r_similarStatementSubmitters[' + index + '][postalCode]'"
+            :value="entry.submitterPostalCode">
 
-            <input
-              type="hidden"
-              :name="'r_similarStatementSubmitters[' + index + '][emailAddress]'"
-              :value="entry.submitterEmailAddress">
-          </template>
+          <input
+            type="hidden"
+            :name="'r_similarStatementSubmitters[' + index + '][emailAddress]'"
+            :value="entry.submitterEmailAddress">
+        </template>
 
-          <li
-            v-if="entry.submitterName"
-            class="o-list__item"
-            v-text="entry.submitterName" />
-          <li
-            v-if="entry.submitterEmailAddress"
-            class="o-list__item"
-            v-text="entry.submitterEmailAddress" />
-          <li
-            v-if="entry.submitterAddress"
-            class="o-list__item"
-            v-text="entry.submitterAddress" />
-          <li
-            v-if="entry.submitterHouseNumber"
-            class="o-list__item"
-            v-text="entry.submitterHouseNumber" />
-          <li
-            v-if="entry.submitterPostalCode"
-            class="o-list__item"
-            v-text="entry.submitterPostalCode" />
-          <li
-            v-if="entry.submitterCity"
-            class="o-list__item"
-            v-text="entry.submitterCity" />
-        </ul>
+        <span
+          v-if="entry.submitterName"
+          class="o-list__item separated"
+          v-text="entry.submitterName" />
+        <span
+          v-if="entry.submitterEmailAddress"
+          class="o-list__item separated"
+          v-text="entry.submitterEmailAddress" />
+        <span
+          v-if="entry.submitterAddress"
+          class="o-list__item separated"
+          v-text="entry.submitterAddress" />
+        <span
+          v-if="entry.submitterHouseNumber"
+          class="o-list__item separated"
+          v-text="entry.submitterHouseNumber" />
+        <span
+          v-if="entry.submitterPostalCode"
+          class="o-list__item separated"
+          v-text="entry.submitterPostalCode" />
+        <span
+          v-if="entry.submitterCity"
+          class="o-list__item separated"
+          v-text="entry.submitterCity" />
       </template>
 
       <template v-slot:form>
         <div
-          data-dp-validate="similarStatementSubmitterForm"
-          class="space-stack-s space-inset-s border">
-          <div :class="fieldsFullWidth ? 'space-stack-s' : 'layout'">
-            <div
-              :class="{ 'layout__item u-1-of-2': !fieldsFullWidth }"
-              class="space-stack-s">
-              <dp-input
-                id="statementSubmitterName"
-                data-cy="voterUsername"
-                v-model="formFields.submitterName"
-                :label="{
-                  text: Translator.trans('name')
-                }"
-                required />
-              <dp-input
-                id="statementSubmitterEmail"
-                data-cy="voterEmail"
-                v-model="formFields.submitterEmailAddress"
-                :label="{
-                  text: Translator.trans('email')
-                }"
-                type="email" />
-            </div><!--
+          class="grid grid-cols-1 gap-x-4"
+          :class="fieldsFullWidth ? '' : 'md:grid-cols-2'"
+          data-dp-validate="similarStatementSubmitterForm">
+          <dp-input
+            id="statementSubmitterName"
+            v-model="formFields.submitterName"
+            class="mb-2"
+            data-cy="voterUsername"
+            :label="{
+              text: Translator.trans('name')
+            }"
+            required />
+          <dp-input
+            id="statementSubmitterEmail"
+            v-model="formFields.submitterEmailAddress"
+            class="mb-2"
+            data-cy="voterEmail"
+            :label="{
+              text: Translator.trans('email')
+            }"
+            type="email" />
 
-         --><div
-              :class="{ 'layout__item u-1-of-2': !fieldsFullWidth }"
-              class="space-stack-s">
-              <div class="o-form__group">
-                <dp-input
-                  id="statementSubmitterAddress"
-                  data-cy="voterStreet"
-                  v-model="formFields.submitterAddress"
-                  class="o-form__group-item"
-                  :label="{
-                    text: Translator.trans('street')
-                  }" />
-                <dp-input
-                  id="statementSubmitterHouseNumber"
-                  data-cy="voterHousenumber"
-                  v-model="formFields.submitterHouseNumber"
-                  class="o-form__group-item shrink"
-                  :label="{
-                    text: Translator.trans('street.number.short')
-                  }"
-                  :size="3" />
-              </div>
+          <div class="o-form__group mb-2">
+            <dp-input
+              id="statementSubmitterAddress"
+              v-model="formFields.submitterAddress"
+              class="o-form__group-item"
+              data-cy="voterStreet"
+              :label="{
+                text: Translator.trans('street')
+              }" />
+            <dp-input
+              id="statementSubmitterHouseNumber"
+              v-model="formFields.submitterHouseNumber"
+              class="o-form__group-item shrink"
+              data-cy="voterHousenumber"
+              :label="{
+                text: Translator.trans('street.number.short')
+              }"
+              :size="3" />
+          </div>
 
-              <div class="o-form__group">
-                <dp-input
-                  id="statementSubmitterPostalCode"
-                  data-cy="voterPostalCode"
-                  v-model="formFields.submitterPostalCode"
-                  class="o-form__group-item shrink"
-                  :label="{
-                    text: Translator.trans('postalcode')
-                  }"
-                  pattern="^[0-9]{4,5}$"
-                  :size="5" />
-                <dp-input
-                  id="statementSubmitterCity"
-                  data-cy="voterCity"
-                  v-model="formFields.submitterCity"
-                  class="o-form__group-item"
-                  :label="{
-                    text: Translator.trans('city')
-                  }" />
-              </div>
-            </div>
+          <div class="o-form__group mb-2">
+            <dp-input
+              id="statementSubmitterPostalCode"
+              v-model="formFields.submitterPostalCode"
+              class="o-form__group-item shrink"
+              data-cy="voterPostalCode"
+              :label="{
+                text: Translator.trans('postalcode')
+              }"
+              pattern="^[0-9]{4,5}$"
+              :size="5" />
+            <dp-input
+              id="statementSubmitterCity"
+              v-model="formFields.submitterCity"
+              class="o-form__group-item"
+              data-cy="voterCity"
+              :label="{
+                text: Translator.trans('city')
+              }" />
           </div>
         </div>
       </template>
@@ -164,14 +156,13 @@
 
 <script>
 import {
-  checkResponse,
   dpApi,
   DpContextualHelp,
   DpEditableList,
   DpInput,
   dpValidateMixin
 } from '@demos-europe/demosplan-ui'
-import { mapMutations } from 'vuex'
+import { mapMutations, mapState } from 'vuex'
 
 export default {
   name: 'SimilarStatementSubmitters',
@@ -238,6 +229,10 @@ export default {
   },
 
   computed: {
+    ...mapState('Statement', {
+      statements: 'items'
+    }),
+
     /**
      * The "add" button text of EditableList is too long when inside the narrow context.
      * This is why a shorter button text is rendered there.
@@ -248,7 +243,7 @@ export default {
         new: Translator.trans('add'),
         add: Translator.trans(this.fieldsFullWidth ? 'add' : 'statement.similarStatementSubmitters.add'),
         abort: Translator.trans('abort'),
-        update: Translator.trans('save'),
+        update: Translator.trans('edit'),
         noEntries: Translator.trans('none'),
         delete: Translator.trans('delete')
       }
@@ -257,7 +252,8 @@ export default {
 
   methods: {
     ...mapMutations('Statement', {
-      updateStatement: 'update'
+      updateStatement: 'update',
+      setInitialStatement: 'setItem'
     }),
 
     ...mapMutations('SimilarStatementSubmitter', {
@@ -302,6 +298,7 @@ export default {
               type: 'SimilarStatementSubmitter'
             }
           })
+          this.setInitialStatement(this.statements[this.statementId])
 
           // Update local state - similarStatementSubmitter
           this.setSimilarStatementSubmitter({
@@ -312,24 +309,29 @@ export default {
     },
 
     deleteEntry (index) {
-      this.updateStatement({
-        id: this.statementId,
-        relationship: 'similarStatementSubmitters',
-        action: 'remove',
-        value: {
-          id: this.listEntries[index].id,
-          type: 'SimilarStatementSubmitter'
+      const name = this.listEntries[index]?.submitterName ? this.listEntries[index].submitterName : false
+
+      if (dpconfirm(Translator.trans('statement.similarStatementSubmitters.delete', { name }))) {
+        this.updateStatement({
+          id: this.statementId,
+          relationship: 'similarStatementSubmitters',
+          action: 'remove',
+          value: {
+            id: this.listEntries[index].id,
+            type: 'SimilarStatementSubmitter'
+          }
+        })
+        this.setInitialStatement(this.statements[this.statementId])
+
+        this.listEntries.splice(index, 1)
+
+        if (this.isRequestFormPost === false) {
+          this.deleteSimilarStatementSubmitter()
         }
-      })
 
-      this.listEntries.splice(index, 1)
-
-      if (this.isRequestFormPost === false) {
-        this.deleteSimilarStatementSubmitter()
-      }
-
-      if (this.isRequestFormPost) {
-        this.resetFormFields()
+        if (this.isRequestFormPost) {
+          this.resetFormFields()
+        }
       }
     },
 
@@ -345,7 +347,6 @@ export default {
       }
 
       dpApi.patch(Routing.generate('api_resource_update', { resourceType: 'Statement', resourceId: this.statementId }), {}, { data: payload })
-        .then(response => { checkResponse(response) })
         .then(() => {
           dplan.notify.notify('confirm', Translator.trans('confirm.entry.deleted'))
         })
@@ -438,7 +439,6 @@ export default {
       }
 
       dpApi.patch(Routing.generate('api_resource_update', { resourceType: 'SimilarStatementSubmitter', resourceId: this.listEntries[index].id }), {}, { data: payload })
-        .then(response => { checkResponse(response) })
         .then(() => {
           // Update local state - similarStatementSubmitter.
           this.setSimilarStatementSubmitter(payload)
